ai-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
anthropic==0.7.1
aiohttp==3.9.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dateutil==2.8.2
yfinance==0.2.32
ta==0.11.0
